Collapse Native/Non-Native Component Tests
We have access to Qunit modules to make this a bit cleaner
